backend/requirements.txt
Flask==2.3.3
requests==2.31.0
gTTS==2.5.1
flask-cors
python-dotenv